declear createPage, updatePage, saveAndReload on page-operation-block
